Remove duplicates and add more type
